+++ packages/platform-core/src/repositories/__tests__/shops.server.test.ts
@@ -1,324 +1,90 @@
-jest.mock("../../dataRoot", () => ({
-  DATA_ROOT: "/data/root",
-}));
+import { jest } from "@jest/globals";
 
-jest.mock("fs", () => ({
-  promises: {
-    readFile: jest.fn(),
+let prismaImportCount = 0;
+let jsonImportCount = 0;
+
+const mockPrisma = {
+  applyThemeData: jest.fn(),
+  readShop: jest.fn(),
+  writeShop: jest.fn(),
+};
+
+const mockJson = {
+  applyThemeData: jest.fn(),
+  readShop: jest.fn(),
+  writeShop: jest.fn(),
+};
+
+jest.mock("../shops.prisma.server", () => {
+  prismaImportCount++;
+  return mockPrisma;
+});
+
+jest.mock("../shops.json.server", () => {
+  jsonImportCount++;
+  return mockJson;
+});
+
+jest.mock("../../db", () => ({ prisma: { shop: {} } }));
+
+const resolveRepoMock = jest.fn(
+  async (_delegate: any, prismaModule: any, jsonModule: any) => {
+    if (process.env.INVENTORY_BACKEND === "json") {
+      return await jsonModule();
+    }
+    return await prismaModule();
   },
-}));
+);
 
-jest.mock("../../db", () => ({
-  prisma: {
-    shop: {
-      findUnique: jest.fn(),
-    },
-  },
-}));
+jest.mock("../repoResolver", () => ({ resolveRepo: resolveRepoMock }));
 
-jest.mock("../../themeTokens/index", () => ({
-  baseTokens: { base: "base" },
-  loadThemeTokens: jest.fn(async () => ({ theme: "theme" })),
-}));
-
-jest.mock("../shop.server", () => ({
-  updateShopInRepo: jest.fn(async (_shop: string, patch: any) => patch),
-}));
-
-import { promises as fs } from "fs";
-import { prisma } from "../../db";
-import { shopSchema } from "@acme/types";
-import * as shops from "../shops.server";
-import { updateShopInRepo } from "../shop.server";
-import { loadThemeTokens } from "../../themeTokens/index";
-
-const { readShop, writeShop } = shops;
-
-describe("shops repository", () => {
-  const findUnique = prisma.shop.findUnique as jest.Mock;
-  const readFile = fs.readFile as jest.Mock;
-  const updateRepo = updateShopInRepo as jest.Mock;
-  const loadTokens = loadThemeTokens as jest.Mock;
+describe("shops repository backend selection", () => {
+  const origBackend = process.env.INVENTORY_BACKEND;
+  const origDbUrl = process.env.DATABASE_URL;
 
   beforeEach(() => {
+    jest.resetModules();
     jest.clearAllMocks();
+    prismaImportCount = 0;
+    jsonImportCount = 0;
+    process.env.DATABASE_URL = "postgres://test";
   });
 
-  describe("readShop", () => {
-    it("returns shop from Prisma when available", async () => {
-      const dbData = {
-        id: "db-shop",
-        name: "DB Shop",
-        catalogFilters: [],
-        themeId: "base",
-        filterMappings: {},
-        themeDefaults: { color: "green" },
-        themeOverrides: { color: "blue" },
-      };
-      findUnique.mockResolvedValue({ data: dbData });
-
-      const result = await readShop("db-shop");
-
-      expect(result.name).toBe("DB Shop");
-      expect(result.themeDefaults).toEqual({ color: "green" });
-      expect(result.themeTokens).toEqual({ color: "blue" });
-      expect(readFile).not.toHaveBeenCalled();
-      expect(loadTokens).not.toHaveBeenCalled();
-    });
-
-    it("falls back to filesystem when Prisma fails", async () => {
-      findUnique.mockRejectedValue(new Error("db fail"));
-      const fileData = {
-        id: "shop1",
-        name: "FS Shop",
-        catalogFilters: [],
-        themeId: "base",
-        filterMappings: {},
-        themeDefaults: { color: "red" },
-        themeOverrides: { color: "blue" },
-      };
-      readFile.mockResolvedValue(JSON.stringify(fileData));
-
-      const result = await readShop("shop1");
-
-      expect(result.name).toBe("FS Shop");
-      expect(result.themeDefaults).toEqual({ color: "red" });
-      expect(result.themeOverrides).toEqual({ color: "blue" });
-      expect(result.themeTokens).toEqual({ color: "blue" });
-      expect(findUnique).toHaveBeenCalled();
-      expect(readFile).toHaveBeenCalledWith(
-        "/data/root/shop1/shop.json",
-        "utf8",
-      );
-      expect(loadTokens).not.toHaveBeenCalled();
-    });
-
-    it("falls back to filesystem when Prisma returns null", async () => {
-      findUnique.mockResolvedValue(null);
-      const fileData = {
-        id: "shop-null",
-        name: "FS Null",
-        catalogFilters: [],
-        themeId: "base",
-        filterMappings: {},
-        themeDefaults: { color: "red" },
-        themeOverrides: { color: "blue" },
-      };
-      readFile.mockResolvedValue(JSON.stringify(fileData));
-
-      const result = await readShop("shop-null");
-
-      expect(result.name).toBe("FS Null");
-      expect(findUnique).toHaveBeenCalled();
-      expect(readFile).toHaveBeenCalledWith(
-        "/data/root/shop-null/shop.json",
-        "utf8",
-      );
-      expect(result.themeTokens).toEqual({ color: "blue" });
-    });
-
-    it("falls back to filesystem when Prisma returns invalid data", async () => {
-      const badDbData = {
-        id: "shop-bad",
-        name: 123,
-        catalogFilters: [],
-        themeId: "base",
-        filterMappings: {},
-      } as any;
-      findUnique.mockResolvedValue({ data: badDbData });
-      const fileData = {
-        id: "shop-bad",
-        name: "FS Fallback",
-        catalogFilters: [],
-        themeId: "base",
-        filterMappings: {},
-        themeDefaults: { color: "red" },
-        themeOverrides: { color: "blue" },
-      };
-      readFile.mockResolvedValue(JSON.stringify(fileData));
-
-      const result = await readShop("shop-bad");
-
-      expect(result.name).toBe("FS Fallback");
-      expect(readFile).toHaveBeenCalledWith(
-        "/data/root/shop-bad/shop.json",
-        "utf8",
-      );
-      expect(loadTokens).not.toHaveBeenCalled();
-    });
-
-    it("returns empty shop with defaults when db returns null and fs fails", async () => {
-      findUnique.mockResolvedValue(null);
-      readFile.mockRejectedValue(new Error("fs fail"));
-
-      const result = await readShop("missing");
-
-      expect(result.id).toBe("missing");
-      expect(result.themeDefaults).toEqual({ base: "base", theme: "theme" });
-      expect(result.themeOverrides).toEqual({});
-      expect(result.themeTokens).toEqual({ base: "base", theme: "theme" });
-      expect(result.analyticsEnabled).toBe(false);
-      expect(result.subscriptionsEnabled).toBe(false);
-      expect(loadTokens).toHaveBeenCalled();
-    });
-
-    it("returns empty shop with defaults when fs has invalid data", async () => {
-      findUnique.mockRejectedValue(new Error("db fail"));
-      const invalidFileData = {
-        name: "No ID",
-        catalogFilters: [],
-        themeId: "base",
-        filterMappings: {},
-      };
-      readFile.mockResolvedValue(JSON.stringify(invalidFileData));
-
-      const result = await readShop("broken");
-
-      expect(result.id).toBe("broken");
-      expect(result.themeDefaults).toEqual({ base: "base", theme: "theme" });
-      expect(result.themeOverrides).toEqual({});
-      expect(result.themeTokens).toEqual({ base: "base", theme: "theme" });
-      expect(loadTokens).toHaveBeenCalled();
-    });
-
-    it("returns empty shop with defaults when fs contains invalid JSON", async () => {
-      findUnique.mockResolvedValue(null);
-      readFile.mockResolvedValue("{bad-json");
-
-      const result = await readShop("invalid-json");
-
-      expect(result.id).toBe("invalid-json");
-      expect(result.themeDefaults).toEqual({ base: "base", theme: "theme" });
-      expect(result.themeOverrides).toEqual({});
-      expect(result.themeTokens).toEqual({ base: "base", theme: "theme" });
-      expect(loadTokens).toHaveBeenCalled();
-    });
-
-    it("uses default tokens when themeDefaults is empty", async () => {
-      findUnique.mockRejectedValue(new Error("db fail"));
-      const fileData = {
-        id: "shop2",
-        name: "No Defaults",
-        catalogFilters: [],
-        themeId: "other",
-        themeDefaults: {},
-        filterMappings: {},
-      };
-      readFile.mockResolvedValue(JSON.stringify(fileData));
-
-      const result = await readShop("shop2");
-
-      expect(result.themeDefaults).toEqual({ base: "base", theme: "theme" });
-      expect(loadTokens).toHaveBeenCalledWith("other");
-    });
-
-    it("loads theme tokens when defaults are missing", async () => {
-      findUnique.mockResolvedValue({
-        data: {
-          id: "shop-no-defaults",
-          name: "No Defaults",
-          catalogFilters: [],
-          themeId: "base",
-          filterMappings: {},
-          themeOverrides: { color: "blue" },
-        },
-      });
-
-      const result = await readShop("shop-no-defaults");
-
-      expect(result.themeDefaults).toEqual({ base: "base", theme: "theme" });
-      expect(result.themeOverrides).toEqual({ color: "blue" });
-      expect(result.themeTokens).toEqual({
-        base: "base",
-        theme: "theme",
-        color: "blue",
-      });
-      expect(readFile).not.toHaveBeenCalled();
-      expect(loadTokens).toHaveBeenCalledWith("base");
-    });
-
-    it("sets empty overrides when overrides are missing", async () => {
-      findUnique.mockResolvedValue({
-        data: {
-          id: "shop-no-overrides",
-          name: "No Overrides",
-          catalogFilters: [],
-          themeId: "base",
-          filterMappings: {},
-          themeDefaults: { color: "green" },
-        },
-      });
-
-      const result = await readShop("shop-no-overrides");
-
-      expect(result.themeDefaults).toEqual({ color: "green" });
-      expect(result.themeOverrides).toEqual({});
-      expect(result.themeTokens).toEqual({ color: "green" });
-      expect(readFile).not.toHaveBeenCalled();
-      expect(loadTokens).not.toHaveBeenCalled();
-    });
+  afterEach(() => {
+    if (origBackend === undefined) {
+      delete process.env.INVENTORY_BACKEND;
+    } else {
+      process.env.INVENTORY_BACKEND = origBackend;
+    }
+    if (origDbUrl === undefined) {
+      delete process.env.DATABASE_URL;
+    } else {
+      process.env.DATABASE_URL = origDbUrl;
+    }
   });
 
-  describe("writeShop", () => {
-    it("merges theme data and prunes overrides", async () => {
-      const current = {
-        ...shopSchema.parse({
-          id: "shop1",
-          name: "Shop",
-          catalogFilters: [],
-          themeId: "base",
-          filterMappings: {},
-          themeDefaults: { color: "red" },
-          themeOverrides: { color: "blue", spacing: "10px" },
-        }),
-      } as any;
+  it("uses Prisma backend by default", async () => {
+    delete process.env.INVENTORY_BACKEND;
+    const repo = await import("../shops.server");
 
-      const readSpy = jest
-        .spyOn(shops, "readShop")
-        .mockResolvedValue(current);
+    await repo.readShop("shop1");
 
-      const patch = {
-        id: "shop1",
-        themeDefaults: { spacing: "10px", extraDefault: "value" },
-        themeOverrides: {
-          color: null,
-          spacing: "10px",
-          extraDefault: "value",
-          newOverride: "15px",
-        } as any,
-      };
+    expect(prismaImportCount).toBe(1);
+    expect(jsonImportCount).toBe(0);
+    expect(mockPrisma.readShop).toHaveBeenCalled();
+    expect(resolveRepoMock).toHaveBeenCalledTimes(1);
+  });
 
-      const result = await writeShop("shop1", patch);
+  it("uses JSON backend when INVENTORY_BACKEND=json", async () => {
+    process.env.INVENTORY_BACKEND = "json";
+    const repo = await import("../shops.server");
 
-      expect(updateRepo).toHaveBeenCalledWith(
-        "shop1",
-        expect.objectContaining({
-          id: "shop1",
-          themeDefaults: {
-            color: "red",
-            spacing: "10px",
-            extraDefault: "value",
-          },
-          themeOverrides: { newOverride: "15px" },
-          themeTokens: {
-            color: "red",
-            spacing: "10px",
-            extraDefault: "value",
-            newOverride: "15px",
-          },
-        }),
-      );
+    await repo.readShop("shop1");
 
-      expect(result.themeOverrides).toEqual({ newOverride: "15px" });
-      expect(result.themeTokens).toEqual({
-        color: "red",
-        spacing: "10px",
-        extraDefault: "value",
-        newOverride: "15px",
-      });
-
-      readSpy.mockRestore();
-    });
+    expect(jsonImportCount).toBe(1);
+    expect(prismaImportCount).toBe(0);
+    expect(mockJson.readShop).toHaveBeenCalled();
+    expect(resolveRepoMock).toHaveBeenCalledTimes(1);
   });
 });
