+++ packages/platform-core/src/repositories/shop.json.server.ts
@@ -1,0 +1,48 @@
+// packages/platform-core/src/repositories/shop.json.server.ts
+import "server-only";
+
+// Import from "fs" so that test mocks can intercept the calls.
+import { promises as fs } from "fs";
+import * as path from "path";
+
+import { shopSchema, type Shop } from "@acme/types";
+import { validateShopName } from "../shops/index";
+import { DATA_ROOT } from "../dataRoot";
+
+function shopPath(shop: string): string {
+  shop = validateShopName(shop);
+  return path.join(DATA_ROOT, shop, "shop.json");
+}
+
+async function ensureDir(shop: string): Promise<void> {
+  shop = validateShopName(shop);
+  await fs.mkdir(path.join(DATA_ROOT, shop), { recursive: true });
+}
+
+export async function getShopById<T extends Shop>(shop: string): Promise<T> {
+  try {
+    const buf = await fs.readFile(shopPath(shop), "utf8");
+    const data = JSON.parse(buf);
+    const parsed = shopSchema.parse(data);
+    return parsed as T;
+  } catch {
+    throw new Error(`Shop ${shop} not found`);
+  }
+}
+
+export async function updateShopInRepo<T extends Shop>(
+  shop: string,
+  patch: Partial<T> & { id: string },
+): Promise<T> {
+  const current = await getShopById<T>(shop);
+  if (current.id !== patch.id) {
+    throw new Error(`Shop ${patch.id} not found in ${shop}`);
+  }
+  const updated = shopSchema.parse({ ...current, ...patch }) as T;
+  await ensureDir(shop);
+  const tmp = `${shopPath(shop)}.${Date.now()}.tmp`;
+  await fs.writeFile(tmp, JSON.stringify(updated, null, 2), "utf8");
+  await fs.rename(tmp, shopPath(shop));
+  return updated;
+}
+
